+++ packages/synapse-interface/slices/portfolio/actions.ts
@@ -1,7 +1,4 @@
-import { Address } from 'viem'
 import { createAction } from '@reduxjs/toolkit'
-
-import { Token } from '@/utils/types'
 
 export enum PortfolioTabs {
   PORTFOLIO = 'portfolio',
@@ -21,13 +18,6 @@ export const setActiveTab = createAction<PortfolioTabs>(
 export const typeSearchInput = createAction<{ searchInput: string }>(
   'portfolio/typeSearchInput'
 )
-
-export const updateSingleTokenAllowance = createAction<{
-  chainId: number
-  allowance: bigint
-  spender: Address
-  token: Token
-}>('portfolio/updateSingleTokenAllowance')
 export const resetPortfolioState = createAction<void>(
   'portfolio/resetPortfolioState'
 )