+++ packages/sdk-router/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.17](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.3.16...@synapsecns/sdk-router@0.3.17) (2024-01-11)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
+## [0.3.16](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.3.15...@synapsecns/sdk-router@0.3.16) (2024-01-11)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
+## [0.3.15](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.3.14...@synapsecns/sdk-router@0.3.15) (2024-01-10)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
+## [0.3.14](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.3.13...@synapsecns/sdk-router@0.3.14) (2024-01-10)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
+## [0.3.13](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.3.12...@synapsecns/sdk-router@0.3.13) (2024-01-10)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
+## [0.3.12](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.3.11...@synapsecns/sdk-router@0.3.12) (2024-01-10)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
+## [0.3.11](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.3.10...@synapsecns/sdk-router@0.3.11) (2024-01-10)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
 ## [0.3.10](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.3.9...@synapsecns/sdk-router@0.3.10) (2024-01-09)
 
 **Note:** Version bump only for package @synapsecns/sdk-router